+++ plugin.image.mypicsdb3/resources/lib/mypicsdb3/__init__.py
@@ -1,5 +1,5 @@
 """Core package for MyPicsDB 3."""
 
 ADDON_ID = "plugin.image.mypicsdb3"
-VERSION = "0.3.5"
+VERSION = "0.4.0"
 SCHEMA_VERSION = 5